Refactor getDefaultVisibility to utils module
Moved the getDefaultVisibility function from plugin.js and test file into a new utils.js module for better reusability and testability. Updated imports and usage in both the plugin and its tests to use the new utility function.
